Training and Test Dataset
Advertising Data which consists of the sales of a product in 200 different markets, along with advertising budgets for the product in each of those markets for three different media: TV, radio, and newspaper. The first 190 rows of the data are used as training set and 10 remaining rows are used as test set.